refactor(DC-704): group settings validators into Configuration/Validators folder
