+++ apps/portal/src/SEBT.Portal.Infrastructure/Configuration/Validators/OidcStepUpSettingsValidator.cs
@@ -2,7 +2,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
 using SEBT.Portal.Core.AppSettings;
 
-namespace SEBT.Portal.Infrastructure.Configuration;
+namespace SEBT.Portal.Infrastructure.Configuration.Validators;
 
 /// <summary>
 /// When any step-up value is set, requires a complete step-up client (and a redirect via step-up or primary callback URI).
