+++ apps/portal/src/SEBT.Portal.Infrastructure/Configuration/Validators/OutageScheduleSettingsValidator.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Options;
 using SEBT.Portal.Core.AppSettings;
 
-namespace SEBT.Portal.Infrastructure.Configuration;
+namespace SEBT.Portal.Infrastructure.Configuration.Validators;
 
 /// <summary>
 /// Rejects a malformed outage schedule at startup rather than letting the evaluator skip windows it
