+++ apps/portal/src/SEBT.Portal.Infrastructure/Configuration/Validators/PiiEncryptionSettingsValidator.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Options;
 using SEBT.Portal.Core.AppSettings;
 
-namespace SEBT.Portal.Infrastructure.Configuration;
+namespace SEBT.Portal.Infrastructure.Configuration.Validators;
 
 /// <summary>
 /// Ensures <see cref="PiiEncryptionSettings"/> binds coherently: key ring parses, 256-bit key lengths, and ActiveKeyId resolves.
