+++ apps/portal/test/SEBT.Portal.Tests/Unit/Configuration/IdProofingRequirementsCoherenceValidatorTests.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Options;
 using SEBT.Portal.Core.AppSettings;
 using SEBT.Portal.Core.Models.Auth;
 using SEBT.Portal.Infrastructure.Configuration;
+using SEBT.Portal.Infrastructure.Configuration.Validators;
 
 namespace SEBT.Portal.Tests.Unit.Configuration;
