+++ apps/portal/test/SEBT.Portal.Tests/Unit/Configuration/PiiEncryptionSettingsValidatorTests.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.Options;
 using SEBT.Portal.Core.AppSettings;
 using SEBT.Portal.Infrastructure.Configuration;
+using SEBT.Portal.Infrastructure.Configuration.Validators;
 using Xunit;
 
 namespace SEBT.Portal.Tests.Unit.Configuration;
