+++ apps/portal/test/SEBT.Portal.Tests/Unit/Configuration/SelfServiceRulesSettingsValidatorTests.cs
@@ -2,6 +2,7 @@ using Microsoft.Extensions.Configuration;
 using SEBT.Portal.Core.AppSettings;
 using SEBT.Portal.Core.Models.Household;
 using SEBT.Portal.Infrastructure.Configuration;
+using SEBT.Portal.Infrastructure.Configuration.Validators;
 
 namespace SEBT.Portal.Tests.Unit.Configuration;
