+++ apps/portal/test/SEBT.Portal.Tests/Unit/Configuration/SocureSettingsValidatorTests.cs
@@ -2,6 +2,7 @@ using Microsoft.Extensions.Hosting;
 using NSubstitute;
 using SEBT.Portal.Core.AppSettings;
 using SEBT.Portal.Infrastructure.Configuration;
+using SEBT.Portal.Infrastructure.Configuration.Validators;
 
 namespace SEBT.Portal.Tests.Unit.Configuration;
